Widen the number pipeline to u128 (2100-kei u64 wraparound fix)
parse_number computed left * 10^power in u64. 2100-kei = 2.1e19
exceeds u64::MAX, wrapping in release builds to 2,553,255,926,
290,448,384 - which then romanized faithfully as the wrong number
(corpus row idxs=[664844]) - and panicking in debug builds. Upstream
numbers.lisp computes in bignums.

The widening is exact, not approximate: both engines gate the counter
path to number texts of at most 20 characters (dict.lisp:1095-1098,
ported in dict/path.rs), so the largest reachable value is 19 digits
+ kei, about 10^35, inside u128 with ~3400x headroom. parse_number
uses checked arithmetic and reports values past u128 as NotANumber,
which find_counter absorbs the same way upstream's not-a-number
handler does: the counter candidate is dropped, the text still
segments, and no error reaches the user. Oversized digit runs never
reach the parser and pass through verbatim, byte-identical to live
ichiran on the same input.

Widened end to end: parse_number / parse_number_star_ (now checked,
Option), number_to_kanji, number_to_kana, CounterText.number,
counter_join, ordinal_str, get_digit, digit_kana_char_len, and the
audit fixture decoder. The truncating "as i64" casts in the counter
kana tables are gone. ordinal_str's domain is unsigned now (every
caller feeds parse_number output), so the negative-input test pinning
Lisp floor-mod went with it. MAX_FIRST_FAILURES raised 10 -> 1000 so
complete burn-in fail sets land in the log.

New unit tests pin REPL-probed values: the corpus 2100-kei row
(nisen hyakkeinen), the stacked-kei window at 10^33, the largest
gated value (19 nines + kei), and the u128-overflow refusal.
docs/known_issues.md gains the conjugation base-form display pick
section (the yoi-vs-ii DB-ordering class root-caused this session).

Gates: suite 1281/0/3; 2k smoke 2000/2000; 252k pass=252310 fail=32,
fail-row set identical to baseline; 1.5M pass=1496900 fail=363 with
the 2100-kei row flipping to pass and every other fail row identical.

+++ kaniran-core/src/dict/counters/classes/tests.rs
@@ -141,7 +141,7 @@ fn ordinal_recipe_propagates_to_base() {
 }
 
 // --- counter_hifumi_class ---
-fn make_ct(number: u64, kana: &str) -> CounterText {
+fn make_ct(number: u128, kana: &str) -> CounterText {
     CounterText {
         text: kana.to_string(),
         kana: kana.to_string(),
